Updated CommentController.java to use CommentRepository and return 404 when not finding comment

+++ backend/src/main/java/no/kristiania/munchMoment/database/CommentRepository.java
@@ -3,5 +3,5 @@ package no.kristiania.munchMoment.database;
 import no.kristiania.munchMoment.entities.Comment;
 import org.springframework.data.repository.ListCrudRepository;
 
-public interface CommentRepository extends ListCrudRepository<Comment,Integer> {
+public interface CommentRepository extends ListCrudRepository<Comment,Long> {
 }
